Move state package to core directory
State is a core data structure needed to maintain starknet's data
